backend/requirements.txt
fastapi
uvicorn
pydantic
loguru
httpx
python-dotenv
requests
langchain-community
dashscope